Update _open-source.scss
Replaced the class 'link' with the attribute <a> so that the hover style applies without requiring the 'class="link"'. When the svg is modified on Illustrator, we lose the class. To get the hover effect, we would need to add 'class=link' to each <a> tag. Cumbersome. The style will only apply on the Open Source page.